:recycles: Moved interfaces
