+++ mastodon_bot/mastodon_bot/__init__.py
@@ -1,2 +1,1 @@
-from .mastodon_bot import BotInterface  # noqa
-from .mastodon_bot import MastodonBot  # noqa
+from .mastodon_bot import MastodonBot
