+++ mastodon_bot/mastodon_bot/interfaces/functions.py
