+++ mastodon_bot/streaming/following.py
@@ -5,8 +5,10 @@ from mastodon import Mastodon
 from mastodon.types import Notification
 from rq_scheduler import Scheduler
 
-from mastodon_bot.mastodon_bot.mastodon_bot import BotInterface
-from mastodon_bot.mastodon_bot.streaming import CallbackStreamListener
+from mastodon_bot.mastodon_bot.interfaces.bot import BotInterface
+from mastodon_bot.mastodon_bot.interfaces.streaming import (
+    CallbackStreamListener,
+)
 
 logger = logging.getLogger(__name__)
